Allow specifying read_timeout (passed through to Net::HTTP)...
There are a few reasons why you might want to do this.  One is that
Net::HTTP's read_timeout is implemented using timeout.rb, which is known
to have race conditions, and in particular is unsafe to use in
EventMachine (because it can sometimes throw exceptions in the main
event loop and kill the whole process); so this lets you disable the
timeout entirely.